LPS-37929 Created RSSRenderer for assetPublisher. Copied for from RSSAction to populateFeedEntries
