+++ portal-impl/src/com/liferay/portlet/assetpublisher/action/RSSAction.java
@@ -78,44 +78,6 @@ public class RSSAction extends com.liferay.portal.struts.RSSAction {
 
 		syndFeed.setEntries(syndEntries);
 
-		for (AssetEntry assetEntry : assetEntries) {
-			SyndEntry syndEntry = new SyndEntryImpl();
-
-			String author = PortalUtil.getUserName(assetEntry);
-
-			syndEntry.setAuthor(author);
-
-			SyndContent syndContent = new SyndContentImpl();
-
-			syndContent.setType(RSSUtil.ENTRY_TYPE_DEFAULT);
-
-			String value = null;
-
-			String languageId = LanguageUtil.getLanguageId(portletRequest);
-
-			if (displayStyle.equals(RSSUtil.DISPLAY_STYLE_TITLE)) {
-				value = StringPool.BLANK;
-			}
-			else {
-				value = assetEntry.getSummary(languageId, true);
-			}
-
-			syndContent.setValue(value);
-
-			syndEntry.setDescription(syndContent);
-
-			String link = getEntryURL(
-				portletRequest, portletResponse, linkBehavior, assetEntry);
-
-			syndEntry.setLink(link);
-
-			syndEntry.setPublishedDate(assetEntry.getPublishDate());
-			syndEntry.setTitle(assetEntry.getTitle(languageId, true));
-			syndEntry.setUpdatedDate(assetEntry.getModifiedDate());
-			syndEntry.setUri(syndEntry.getLink());
-
-			syndEntries.add(syndEntry);
-		}
 
 		syndFeed.setFeedType(RSSUtil.getFeedType(format, version));
 
@@ -145,124 +107,6 @@ public class RSSAction extends com.liferay.portal.struts.RSSAction {
 		syndFeed.setUri(feedURL);
 
 		return RSSUtil.export(syndFeed);
-	}
-
-	protected List<AssetEntry> getAssetEntries(
-			PortletRequest portletRequest,
-			PortletPreferences portletPreferences)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		int rssDelta = GetterUtil.getInteger(
-			portletPreferences.getValue("rssDelta", "20"));
-
-		return AssetPublisherUtil.getAssetEntries(
-			portletPreferences, themeDisplay.getLayout(),
-			themeDisplay.getScopeGroupId(), rssDelta, true);
-	}
-
-	protected String getAssetPublisherURL(PortletRequest portletRequest)
-		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		Layout layout = themeDisplay.getLayout();
-
-		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
-
-		StringBundler sb = new StringBundler(7);
-
-		String layoutFriendlyURL = GetterUtil.getString(
-			PortalUtil.getLayoutFriendlyURL(layout, themeDisplay));
-
-		if (!layoutFriendlyURL.startsWith(Http.HTTP_WITH_SLASH) &&
-			!layoutFriendlyURL.startsWith(Http.HTTPS_WITH_SLASH)) {
-
-			sb.append(themeDisplay.getPortalURL());
-		}
-
-		sb.append(layoutFriendlyURL);
-		sb.append(Portal.FRIENDLY_URL_SEPARATOR);
-		sb.append("asset_publisher/");
-		sb.append(portletDisplay.getInstanceId());
-		sb.append(StringPool.SLASH);
-
-		return sb.toString();
-	}
-
-	protected String getEntryURL(
-			PortletRequest portletRequest, PortletResponse portletResponse,
-			String linkBehavior, AssetEntry assetEntry)
-		throws Exception {
-
-		if (linkBehavior.equals("viewInPortlet")) {
-			return getEntryURLViewInContext(
-				portletRequest, portletResponse, assetEntry);
-		}
-		else {
-			return getEntryURLAssetPublisher(
-				portletRequest, portletResponse, assetEntry);
-		}
-	}
-
-	protected String getEntryURLAssetPublisher(
-			PortletRequest portletRequest, PortletResponse portletResponse,
-			AssetEntry assetEntry)
-		throws Exception {
-
-		AssetRendererFactory assetRendererFactory =
-			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
-				assetEntry.getClassName());
-
-		StringBundler sb = new StringBundler(4);
-
-		sb.append(getAssetPublisherURL(portletRequest));
-		sb.append(assetRendererFactory.getType());
-		sb.append("/id/");
-		sb.append(assetEntry.getEntryId());
-
-		return sb.toString();
-	}
-
-	protected String getEntryURLViewInContext(
-			PortletRequest portletRequest, PortletResponse portletResponse,
-			AssetEntry assetEntry)
-		throws Exception {
-
-		AssetRendererFactory assetRendererFactory =
-			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
-				assetEntry.getClassName());
-
-		AssetRenderer assetRenderer = assetRendererFactory.getAssetRenderer(
-			assetEntry.getClassPK());
-
-		String viewInContextURL = assetRenderer.getURLViewInContext(
-			(LiferayPortletRequest)portletRequest,
-			(LiferayPortletResponse)portletResponse, null);
-
-		if (Validator.isNotNull(viewInContextURL) &&
-			!viewInContextURL.startsWith(Http.HTTP_WITH_SLASH) &&
-			!viewInContextURL.startsWith(Http.HTTPS_WITH_SLASH)) {
-
-			ThemeDisplay themeDisplay =
-				(ThemeDisplay)portletRequest.getAttribute(
-					WebKeys.THEME_DISPLAY);
-
-			viewInContextURL = themeDisplay.getPortalURL() + viewInContextURL;
-		}
-
-		return viewInContextURL;
-	}
-
-	protected String getFeedURL(PortletRequest portletRequest)
-		throws Exception {
-
-		String feedURL = getAssetPublisherURL(portletRequest);
-
-		return feedURL.concat("rss");
 	}
 
 	@Override
